Corretto campo pertcentuale su tabella PERCENTUALEANTICIPOTE da byte a decimal.

+++ NewISE/EF/PERCENTUALEANTICIPOTE.cs
@@ -25,7 +25,7 @@ namespace NewISE.EF
         public decimal IDTIPOANTICIPOTE { get; set; }
         public System.DateTime DATAINIZIOVALIDITA { get; set; }
         public System.DateTime DATAFINEVALIDITA { get; set; }
-        public byte PERCENTUALE { get; set; }
+        public decimal PERCENTUALE { get; set; }
         public System.DateTime DATAAGGIORNAMENTO { get; set; }
         public bool ANNULLATO { get; set; }
